predict/requirements.txt
Flask==3.0.0
keras==2.15.0
numpy==1.26.2
Pillow==9.4.0
Requests==2.31.0
tensorflow==2.15.0
boto3==1.34.8
